requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
coverage>=7.3.2
flake8>=6.1.0
black>=23.11.0
selenium>=4.15.2  # For E2E testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0